LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/script-src-module.tentative-expected.txt

CONSOLE MESSAGE: TypeError: 'text/plain' is not a valid JavaScript MIME type.


PASS Speculative parsing, page load: script-src-module
